chore: resolve linter complaints
