+++ aws-scanners/ecs/tsconfig.json
@@ -1,6 +1,13 @@
 {
   "extends": "tsconfig/base.json",
-  "exclude": ["codegen.ts", "config.ts", "test/*.test.ts"],
+  "include": [
+    "src"
+  ],
+  "exclude": [
+    "codegen.ts",
+    "config.ts",
+    "test/*.test.ts"
+  ],
   "compilerOptions": {
     "target": "ES2015",
     "module": "Node16",
