+++ packages/cytoscape-serializer/src/index.ts
@@ -61,38 +61,34 @@ export type EdgeTarget = {
 };
 
 export function serializeGraph(graph: Graph): CytoscapeGraph[] {
-    const nodes: CytoscapeNode[] = graph.nodes.map((node) => {
-        return {
-            group: "nodes",
+    const nodes: CytoscapeNode[] = graph.nodes.map((node) => ({
+        group: "nodes",
+        id: node.id,
+        data: {
             id: node.id,
-            data: {
-                id: node.id,
-                parent: typeof node.parent === 'string' ? node.parent : node.parent?.id,
-                name: node.name,
-                service: node.service as string,
-                type: node.type ?? node.service ?? 'node'
-            },
-            metadata: node.metadata
-        };
-    });
+            parent: typeof node.parent === 'string' ? node.parent : node.parent?.id,
+            name: node.name,
+            service: node.service as string,
+            type: node.type ?? node.service ?? 'node'
+        },
+        metadata: node.metadata
+    }));
 
-    const edges: CytoscapeEdge[] = graph.edges.map((edge) => {
-        return {
-            group: 'edges',
+    const edges: CytoscapeEdge[] = graph.edges.map((edge) => ({
+        group: 'edges',
+        id: edge.id,
+        data: {
             id: edge.id,
-            data: {
-                id: edge.id,
-                name: edge.name as string,
-                source: edge.source.id,
-                target: edge.target.id,
-                type: "edge",
-            },
-            metadata: {
-                label: edge.target.name,
-                ...edge.metadata
-            }
-        };
-    });
+            name: edge.name as string,
+            source: edge.source.id,
+            target: edge.target.id,
+            type: "edge",
+        },
+        metadata: {
+            label: edge.target.name,
+            ...edge.metadata
+        }
+    }));
 
     return [...nodes, ...edges];
 }
