+++ packages/sdk/src/index.ts
@@ -46,7 +46,9 @@ export interface ServiceNodesMap {
 
 export default class Infrascan {
   private regionalScannerRegistry: ScannerRegistry;
+
   private globalScannerRegistry: ScannerRegistry;
+
   private pluginRegistry: {
     [E in GraphPluginEvents]: Omit<GraphPlugin<E>, "event">[];
   };
@@ -296,15 +298,15 @@ export default class Infrascan {
             });
             const serviceNodes =
               serviceNodeMap[
-                regionalServiceScanner.arnLabel ??
-                  regionalServiceScanner.service.toLowerCase()
+              regionalServiceScanner.arnLabel ??
+              regionalServiceScanner.service.toLowerCase()
               ];
             if (serviceNodes) {
               serviceNodes.push(...regionalServiceNodeIds);
             } else {
               serviceNodeMap[
                 regionalServiceScanner.arnLabel ??
-                  regionalServiceScanner.service.toLowerCase()
+                regionalServiceScanner.service.toLowerCase()
               ] = regionalServiceNodeIds;
             }
